// src/app/shared/models/user.ts
export class User {
    usuario: string;
    nombre: string;
    correo: string;
    fechaNacimiento: string;
    razonSocial?: string;
    cif:string;
    telefono?: string;
    password: string;
    tipoCliente: any;
    alergenos?: string;
    necesidadesAlimentacion?: string;
    pedidoCliente?:any;
  }